refactor: 함수를 하나로 정리
